+++ src/TRegx/CleanRegex/Match/MatchPattern.php
@@ -1,7 +1,9 @@
 <?php
 namespace TRegx\CleanRegex\Match;
 
-use TRegx\CleanRegex\Exception\CleanRegex\InvalidReturnValueException;
+use ArrayIterator;
+use EmptyIterator;
+use Iterator;
 use TRegx\CleanRegex\Exception\CleanRegex\NotMatched\Subject\FirstMatchMessage;
 use TRegx\CleanRegex\Exception\CleanRegex\SubjectNotMatchedException;
 use TRegx\CleanRegex\Internal\Factory\NotMatchedOptionalWorker;
@@ -17,9 +19,7 @@ use TRegx\CleanRegex\Match\ForFirst\NotMatchedOptional;
 use TRegx\CleanRegex\Match\ForFirst\Optional;
 use TRegx\CleanRegex\Match\Offset\MatchOffsetLimit;
 use TRegx\CleanRegex\MatchesPattern;
-use TRegx\SafeRegex\Guard\Arrays;
 use TRegx\SafeRegex\preg;
-use function is_array;
 
 class MatchPattern implements PatternLimit
 {
@@ -134,6 +134,15 @@ class MatchPattern implements PatternLimit
     public function count(): int
     {
         return preg::match_all($this->pattern->pattern, $this->subject);
+    }
+
+    public function iterator(): Iterator
+    {
+        $objects = $this->getMatchObjects();
+        if (empty($objects)) {
+            return new EmptyIterator();
+        }
+        return new ArrayIterator($objects);
     }
 
     /**